refactor: consolidate client mappers, functions, and hooks into utils/
