+++ src/client/Editor/Components/EditorQuestionnaireDetails.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from 'react';
 import { QuestionnaireDetails } from '../../../common/interfaces/surveyInterface';
-import { useAsyncRequestWithThreeParams } from '../../Common/hooks/useAsyncRequest';
+import { useAsyncRequestWithThreeParams } from '../../utils/useAsyncRequest';
 import { getEditorInformation } from '../../api/NodeApi';
 import AsyncContent from '../../Common/components/AsyncContent';
 import EditorContent from './EditorContent';
