+++ src/client/Research/Pages/ResearchHome.tsx
@@ -1,6 +1,6 @@
 import { User } from 'blaise-api-node-client';
 import { Survey } from '../../../common/interfaces/surveyInterface';
-import { useAsyncRequestWithParam } from '../../Common/hooks/useAsyncRequest';
+import { useAsyncRequestWithParam } from '../../utils/useAsyncRequest';
 import { getSurveys } from '../../api/NodeApi';
 import AsyncContent from '../../Common/components/AsyncContent';
 import SurveysList from '../../Common/components/SurveysList';
