+++ src/client/api/NodeApi.ts
@@ -3,10 +3,10 @@ import { User } from 'blaise-api-node-client/lib/cjs/interfaces/user';
 import { Survey } from '../../common/interfaces/surveyInterface';
 import { SupervisorInformation } from '../types/supervisorInterface';
 import { EditorInformation } from '../types/editorInterface';
-import mapEditorInformation from '../Mappers/editorInformaitionMapper';
-import mapSupervisorInformation from '../Mappers/supervisorInformationMapper';
+import mapEditorInformation from '../utils/editorInformaitionMapper';
+import mapSupervisorInformation from '../utils/supervisorInformationMapper';
 import { CaseSummaryDetails } from '../../common/interfaces/caseInterface';
-import mapCasesNotAllocated from '../Mappers/caseAllocationMapper';
+import mapCasesNotAllocated from '../utils/caseAllocationMapper';
 import { AllocationDetails } from '../../common/interfaces/allocationInterface';
 import { getDataFromNode, patchDataToNode } from './AxiosApi';
