+++ src/client/utils/supervisorInformationMapper.ts
