+++ src/test/client/Common/components/DownloadCaseSummaryLink.test.tsx
@@ -5,7 +5,7 @@ import '@testing-library/jest-dom';
 import { vi, type MockInstance } from 'vitest';
 import { DownloadCaseSummaryLink, Props } from '../../../../client/Common/components/DownloadCaseSummaryLink';
 import { getCaseSummary } from '../../../../client/api/NodeApi';
-import mapCaseSummaryText from '../../../../client/Mappers/caseSummaryTextMapper';
+import mapCaseSummaryText from '../../../../client/utils/caseSummaryTextMapper';
 import { caseSummaryDetailsMockObject } from '../../../server/mockObjects/CaseMockObject';
 import { clientLogger } from "../../../../client/logger";
