+++ src/test/client/Mappers/caseAllocationMapper.test.ts
@@ -2,7 +2,7 @@ import { CaseEditInformation } from 'blaise-api-node-client/lib/cjs/interfaces/c
 import CaseOutcome from 'blaise-api-node-client/lib/cjs/enums/caseOutcome';
 import EditedStatus from 'blaise-api-node-client/lib/cjs/enums/editedStatus';
 import Organisation from 'blaise-api-node-client/lib/cjs/enums/organisation';
-import mapAllocationDetails from '../../../client/Mappers/caseAllocationMapper';
+import mapAllocationDetails from '../../../client/utils/caseAllocationMapper';
 import { AllocationDetails } from '../../../common/interfaces/allocationInterface';
 
 describe('Map cases not allocated informaiton', () => {
