+++ src/test/client/Mappers/editorInformationMapper.test.ts
@@ -2,7 +2,7 @@ import { CaseEditInformation } from 'blaise-api-node-client/lib/cjs/interfaces/c
 import CaseOutcome from 'blaise-api-node-client/lib/cjs/enums/caseOutcome';
 import EditedStatus from 'blaise-api-node-client/lib/cjs/enums/editedStatus';
 import Organisation from 'blaise-api-node-client/lib/cjs/enums/organisation';
-import mapEditorInformation from '../../../client/Mappers/editorInformaitionMapper';
+import mapEditorInformation from '../../../client/utils/editorInformaitionMapper';
 import { EditorInformation } from '../../../client/types/editorInterface';
 
 describe('Map editor informaiton', () => {
